Raise ConfigException when the PropertyStatistics cannot be built
We just pass through the arguments of the template as parameters
to the Python object constructor. What could possibly go wrong.

If the constructor throws a TypeError, let's catch and re-raise
as a ConfigException.

This kind of leaves it to the user to figure out, but better
than crashing and let the webserver throw a 4XX.