Reorganize files under a more standard package layout
This paves the way to adding more unit tests in a clear way,
and enabling package test coverage.

- Move sources files under a `integraality` package sub-directory
- Move the test suite file to tests sub-directory
  (and make it a package, so that tests are collected)

We do have to do some pretty ugly things with the PYTHONPATH
because that layout implies in theory that we have to make
the package installable, which is little bit unnecessary here.