Remove method run from PropertyStatistics
The `run` method computes the statistics, makes the
wikitext for it and writes it in the given wikipage.

This was initally used during development but we
later developed the PagesProcessor class to take
care of the page update part. Now, both the CLI
and web-based updates rely on it and the run method
is unused.

This removes the `run` method, and changes the `main` method
to simply output the wikitext returned by `retrieve_and_process_data`.