Fix ordering of OrderedDict in TestParseConfigProperties
`parse_config_properties` returns an OrderedDict, so we need
to construct one as expected result in the unit test.

On Travis-CI, `test_more_properties` was failing because
of mismatching dictionary orders.

This is because, interestingly enough, passing a dict (mapping)
to the OrderedDict constructor will not retain the insertion
order of the source code. Only passing an iterable of
key-value pairs will retain the order.

(This had not come-up locally, as I use a local Python 3.6,
in which this has likely been fixed already).