Fix unit test test_make_stats_for_one_grouping_with_higher_grouping
For triggering the higher-grouping behaviour, we also need
the `self.higher_grouping` property to be set.

Although it appears somewhat redundant with passing a higher_grouping
value to the `make_stats_for_one_grouping` method, there is a usecase
for it: in the event where the higher_grouping was configured, but
where no higher-grouping for a given group could be retrieved, then
we would want to add an empty cell.

- Add text fixture setting `self.higher_grouping`
- Fix the expected value accordingly