Remove useless loop in `_get_count_from_sparql`
`_get_count_from_sparql` issues a COUNT query to the Query Service,
and thus expects a single line result with a single value,
ie a list of one element, that being a dictionary of one entry.
Even if there is no result, the result will be a list of one entry
(with a `{'count': 0}`)

There is no point in starting a for loop and return on the first
iteration, while we can just retrieve the first element.