Fix ordering of OrderedDict in PropertyStatisticsTest
We have been having some flakiness on Travis-CI in Python 3.5,
with tests failing because `Calls not found`.

This is because `assert_has_calls` is ordered by default, and as
explained in 8365cc45, building an OrderedDict using a dict (mapping)
will not retain the insertion order of the source code.
Only passing an iterable of key-value pairs will retain the order.