Refactor data structure for property configuration
Up until now, the configuration for properties
(ie, the columns of the future table) was done
using an OrderedDict, where the key was the property
name (its ID) and the value the title to use (if set).

We want to support more complex configuration
(in particular for qualifier support) so we
need to store this information into a more
robust and extensible data structure.

This introduces a new class, `PropertyConfig`, which
for now just holds the property name and the override
title as class attributes.