Extend unit tests for qualifier support
Qualifier-support was added in 6a055c2d4 with decent unit-tests,
however the more generic unit-tests were not updated for the new feature.

This adds two new-style PropertyConfig to the`PropertyStatisticsTest`
fixtures (which are reused by all other unit test classes),
and update all test assertions accordingly.

This also solves a bug where the `P1/Q2/P3` syntax was not working,
as the `Q2` part was never passed to the methods.