Set default property_threshold to 0
The use case for `property_threshold` is quite thin (if any),
and it can be misconfigured in highly surprising ways
(ending up with a dahsboard full of zeroes).

I am unsure what's the best path forward for that variable.
Should we warn the user when it is set too high
(in particular, set  higher than `grouping_threshold`)?

As a stop-gap measure, let's just set its default value to zero,
as it seems to be the most common use-case.

Ref T224226