Guard against no start template found in make_stats_object_for_page
The method `make_stats_object_for_page` iterates over the templates
present on the page, and builds a `PropertyStatistics` object.

T240312 describes a situation where the start template could not be found.

This is an almost impossible situation, because starting Integraality,
whether via link or via weekly batch update, pre-supposes the presence
of that template. This can only be assumed to be a pywikibot issue,
where a special syntax confuses `templatesWithParams`.

For now, let's raise an Exception and fail gracefully when this happens.

Bug: T240312