Also run Travis CI tests on Python 3.7
We started with testing with Python 3.5 (3b4a376c),
as it is the version used on Toolforge.
Since 74ef4daf, we also test on Python 3.6.

Since December 2019, Python 3.7 is available on Toolforge.

Let's start by running the tests under 3.7, and from there
we can look into moving to 3.7, and deprecate 3.5.