Downgrade and pin zipp library to 1.0.0 for Python 3.5 compatibility
In 10f67fd29 we bumped all dev dependencies to the latest version,
including zipp from 0.6.0 to 2.1.0.

However starting zipp 2.0.0 compatiblity with 3.5 was dropped [1],
while this is still the version we are running in Toolforge,
resulting in failing Travis-CI tests.

Until we move to 3.7 (which is now available on Toolforge),
let's downgrade zipp so we can have passing tests.

[1] See https://github.com/jaraco/zipp/issues/28