Untie PagesProcessor to Wikidata as a Site
The role of PagesProcessor is to read the integraality configuration
from a wiki-page, create a PropertyStatistics object with that config,
get back some wikitext to insert in the original page.

As such, PagesProcessor is not particularly tied to Wikidata:
it just happens that we expect the Property Dashboard template
to be used there.

For T243780, we want to make InteGraality operate on more wikis
than only Wikidata.

This recognizes this decoupling by and paves towards that goal.
Instead of creating an en.wp Site from which to get the data_repository,
create a Site based on a given URL (defaulting to Wikidata for backwards
compatibilit ; and operate on that Site throughout.