Add crontab updating daily the dashboards on Meta
With 0fcb5220, we can now process any site using
inteGraality via the command-line.

This adds a crontab to regularly update the Meta dashboards

Since the FindingGLAMs challenge is running this week,
and given how there are not many dashboards on Meta,
let's update them daily instead of the daily Wikidata.

Bug: T243780
Bug: T239067