Add method get_title to PropertyConfig and use it in the cell template
We need to output in the {{integraality cell}} template a reference
to the column input (ie, property or property/qualifier combo).

Turns out that we were not doing this correctly for the non-simple
case (only property).

The get_title method joins together the PropertyConfig pieces in
the same fashion as the original input.

Bug: T251008