Add partial support for Labels
integraality currently only supports properties
(and properties with qualifiers since 6a055c2).

There is no reason we could not support computing
completion for labels and/or descriptions in a given language.

This implements partial support for Labels.

In `property_statistics` module
 - Add LabelConfig object, akin to PropertyConfig,

 - On `PropertyStatistics` object:
   - Add SPARQL queries methods:
     - get_label_info
     - get_label_info_no_grouping
     - get_totals_for_label
   - Hook-in these methods, switching on `isinstance`, in three places:
     - make_stats_for_no_group
     - make_footer
     - retrieve_and_process_data

In `pages_processor` module:
 - extend `parse_config_properties` to support the Labels syntax
   ("Lxx" with xx being the language code)
 - rename around the `properties` argument into `columns`,
   as we changed that interface in PropertyStatistics but we cannot
   change it in one go in the Wiki template.

This is incomplete, as there is no support for the looking glass.

Bug: T248788