Accept "column" as GET parameter in queries endpoint
As part of extending support to labels and descriptions,
we renamed the whole 'property' lingo into the more generic
'column' throughout most of the code, and in the wiki templates.

However the webserver queries endpoint was not updated,
leading to a regression where the property name was
not passed through, leading to misleading SPARQL queries.

With this we support both "property" and "column",
and rename the variable throughout the app module,
the Jinja template and the unit tests.