Add support for Descriptions in PropertyStatistics
Building upon the refactorings in f0db935, f8d9fdf and f57c060,
adding support for descriptions on top of labels is fairly simple:

 - Add `DescriptionConfig` subclassing `TextConfig`
 - Add one conditional to `parse_config_properties`

The Queries endpoint feature is a bit more annoying, hence the
introduction of the `TEXT_SELECTOR_MAPPING` which will hopefully
go away in a future refactoring.

(the work for the Jinja templates had been done already in b74b35).

Bug: T248788