Guard against invalid page names when making grouping links
Integraality offers the option to link rows using the grouping English label.

This would hard crash if the page would not be a valid title for Pywikibot.

In particular this would happen when 'No value' or "Unknown value"
would be a grouping.

This is not per-se a solution to T273226, but a decent workaround
to the worst of it, and arguably a good thing to have regardless.

Bug: T273226