Add mwparserfromhell to requirements to allow advanced SPAQL constructs
The `selector_sparql` can be made arbitrarily complex, and as such may
contain SPARQL constructs that use curly braces (UNIONs or sub-queries).

This trips pywikibot built-in parser that uses regular expressions

As discussed in T278454, it's a known limitation and the recommended
workaround is to install a more advanced parser such as `mwparserfromhell`:
```
mwparserfromhell is a Python package that provides an easy-to-use
and outrageously powerful parser for MediaWiki wikicode.
```

It's a drop-in replacement, so we only need to add it to the requirements.

See also https://github.com/earwig/mwparserfromhell

Bug: T240312