Use Python3.5 virtual environment in cron jobs
In 9d5e608e, we tried migrating to Python3.7, and thus have a Python3.7
virtual environment in the conventional `www/python/venv/`.

However as outlined in T257942, the py37 cannot be used by the cron jobs
executed on the GridEngine, and thus have an addition py35 virtual environment.

We thus need to instruct the shell scripts used by the cron jobs
to source that virtual environment.

(This had been hand-fixed on the server back in July
but erased by a recent deployment).