Replace PropertyStatistics `columns` attribute with an indexed columns
The `columns` attribute is a list of Column object.

We will need a way to access a given column object − based on the key.

Here the attribute becomes a dictionary column_key →  Column object
(which also allows some simplification in places where we needed the key).

Also do some minor alterations to the unit tests accordingly.