Pass column object to `get_query_for_items_for_property_[posi|nega]tive` instead of column key
We are aiming to move some of the logic inside
`get_query_for_items_for_property_[posi|nega]tive`
to the Column object.

A first step towards that is to have the Column object
as argument to the method, and not merely the key.

For now, we basically convert back to the key internally.

This however requires significant overhaul of the `app` tests.