Switch to new Debian Buster toolforge Grid and Python3.7 everywhere
The previous Tools Grid, running on Strech, was using Python 3.5.

To use it, we forced the use of a Python3.5
virtual environment for cron jobs (eac4d2c5).

The previous Tools Grid is now deprecated and
we need to move the workload to Python 3.7 [1].

We were already using Python3.7 for the webservice
(since 9d5e608e07), so there is not that much to do.

- Remove the eac4d2c5 workaround,
  and let the cron job use the normal venv.

- Connect to a Buster Bastion host in Ansible hosts file

- Add `-release buster` flag to jsub commands in cron jobs

[1] See https://wikitech.wikimedia.org/wiki/News/Toolforge_Stretch_deprecation

Bug: T257942