Remove ww package as python3.5 workaround for f-string
integraality was coded from the beginning using f-strings,
which were only added to Python 3.6.

However, Toolforge only allowed at the time Python 3.5,
and it was still required for Grid Engine jobs.

We thus added in 2abc519e7 a workaround with the `ww` package [1].

Now that Python 3.5 is out of the picture since 8bd793d70
and the move to the Debian Stretch Grid Engine and Python 3.7
we can safely remove the ww package and all the fake f-string syntax.

[1] https://pypi.org/project/ww/