Add basic but real unit tests for `retrieve_data` and `process_data`
After splitting-extract these method in 3a675b6,
we only added the most basic unit tests with empty object.

The idea was that we would add actual tests only after
the major Line/Grouping refactoring.

Let's add the tests now, and alter them later,
so that the refactoring is a bit less hairy.