Pass grouping object to `make_stats_for_one_grouping` instead of its attributes
We are aiming to move some of the logic inside
`make_stats_for_one_grouping` to the Line object(s).

A first step towards that is to have the Grouping object
as argument to the method, and not merely the key.

For now, we basically extract back the attributes.

This however requires significant overhaul of the unit tests.