Retire the `column_data` structure and move data to the grouping objects
The data for the table cells was stored in the dictionary
of OrderedDict `column_data`, stored as object attribute
to the PropertyStatistics object, and used later
in `make_stats_for_one_grouping`.

This structure was reflecting the way we retrieve the data,
as we run a SPARQL query per column.

This moves the data for the cells to the grouping objects,
continuing the refactoring started in 70816729.

- Add `cells` attribute to the `AbstractLine` object.
- Change `retrieve_data` to dispatch the cell data to
  each grouping object
- Retire `column_data` everywhere

This paves the way to move even more logic away from
PropertyStatistics into the Line objects, continueing c81b19c.