Switch Statsd host in Toolforge environment to `cloudmetrics1003`.
Since f45589b, the application is instrumented and sends
metrics to the Wikimedia Cloud Services statsd endpoint.

It was initially sent to cloudmetrics1002,
then switched to cloudmetrics1001 (6de23088f).

Since a few weeks, logs report issues reaching that endpoint [1].

This started on 2022-11-20 according to the otherwise fine Shell logs.
(Not sure exactly when it started for the Python webservice,
as the uwsgi logfile ballooned to 31 million lines,
making precise investigation a tad difficult).

It turns out cloudmetrics1002 (as well as cloudmetrics1001)
were decommissioned on November 16th (see T297444).

It also turns out the pystatsd library plain crashes
if the statsd host is unreachable [2].

This replaces the host with the new cloudmetrics1003.

Bug: T325936

[1] Python traceback:
```
Traceback (most recent call last):
  File "/data/project/integraality/www/python/src/app.py", line 10, in <module>
    from pages_processor import PagesProcessor, ProcessingException
  File "./pages_processor.py", line 16, in <module>
    from property_statistics import PropertyStatistics, QueryException
  File "./property_statistics.py", line 16, in <module>
    from statsd.defaults.env import statsd
  File "/data/project/integraality/www/python/venv/lib/python3.7/site-packages/statsd/defaults/env.py", line 17, in <module>
    maxudpsize=maxudpsize, ipv6=ipv6)
  File "/data/project/integraality/www/python/venv/lib/python3.7/site-packages/statsd/client/udp.py", line 35, in __init__
    host, port, fam, socket.SOCK_DGRAM)[0]
  File "/usr/lib/python3.7/socket.py", line 748, in getaddrinfo
    for res in _socket.getaddrinfo(host, port, family, type, proto, flags):
```

[2] https://github.com/jsocol/pystatsd/issues/130