Correctly handle of higher-grouping when there is no higher-grouping
It can be that the higher-grouping is missing for a grouping,
especially if the feature is creatively used to eg display
the logo of a publisher or the flag of a country.

This resulted in a broken formatting, with a missing cell.

It's sometimes to escape the polysemy of `None`:
for higher-grouping it meant both that it was not
configured at all, or was punctually missing.

This should be resolved in a more elegant way, but alas,
let's make a quick fix by using an empty string for the
second use-case.