Rename `PropertyGrouping` object to `ItemGrouping` and move `YearGrouping`
The 'Line' object concept tree introduced in 0d775cf2
mistakenly called this object `PropertyGrouping`,
but in reality it's not about the source of the grouping
but about its contents: in this case an Item page.

Indeed, the methods in that object are only about
formatting the grouping link in case of ItemPage,
and about formatting a higher-grouping, a concept
that only makes sense for Item groupings.

Also move `YearGrouping` straight under `Grouping`,
which tellingly does not require any change in tests.

The new tree is thus:

AbstractLine
└── Grouping
   ├── NoGroupGrouping
   ├── ItemGrouping
   ├── YearGrouping
   ├── UnknownValueGrouping
   └── TotalsGrouping

(there could/might be a need in the future for
a level above Item and Year, but not for now).