Do not add higher grouping empty column when there is no higher grouping
The fix 7b095a8a had the unintended side-effect that a spurious
higher grouping column was added for the "NoGrouping" and "Totals" lines.

There are two sources for whether we have a higher grouping:
the grouping configuration, and the grouping object
(which may or may not have not one, even if configured,
which is what 7b095a8a intended to fix).

Here we also test whether a higher grouping was configured at all,
and never add a higher grouping column in that case.

Fixes #16