Add support for grouping by any Wikibase predicate
The initial idea of integraality was to group by Property,
or more exactly the (distinct) items used as values of the property.
The assumption was that properties have items for values.
Later on, we introduced grouping for properties of the `time` datatype
(by casting to year).

But properties are not the only Wikibase concept we might want to group along.

In particular for Lexemes, the language and the lexical category make more sense
(respectively `dct:language` and `wikibase:lexicalCategory` predicates).

Here we introduce a new element in the GroupingConfiguration concept tree:
`PredicateGroupingConfiguration` is used every time the 'grouping property'
does not match the Pxyz format.

The assumption is that all predicates have Wikibase items as values,
but that seems to actually be the case.

This opens the door to dashboards for lexemes.

Bug: T312729