Retire the use of `grouping property` for `grouping predicate`
Following the limitation that only properties were supported
as grouping concept, the line objects were taking a property
as parameter and adding a hardcoded `wdt:` prefix.

Now that we support any type of predicate for grouping,
let's have the line objects use that predicate directly.

The GroupingConfiguration objects now have a `get_predicate` method,
and that's where the `wdt:` is added for the `PropertyGroupingConfiguration`.

Also solves a hard crash [1] when using the queries feature
on dashboards using Predicates as grouping.

[1] Traceback
```
AttributeError: 'PredicateGroupingConfiguration' object has no attribute 'property'
```

Bug: T312729