Add support for sitelink column
The sitelink column would display the percentage of items
in the group that have a page on the given wikimedia project.

The syntax used is the "database name", eg `frwiki` or `bnwikivoyage`.
We use a mapping (stored as JSON file) to map to the Qid and URL
of the project − respectively for display and SPARQL querying.

(The mapping was done by querying Wikidata and its P1800 property)

Bug: T312726