Remove special case for NoGroupGrouping when building negative queries
Back in 06cb52d and ac64840b, when moving the SPARQL-formatting code from
`get_query_for_items_for_property_nega|positive` to the Line objects,
we introduced a special case for NoGroupGrouping.

The reason was that both the grouping exclusion and the column exclusion
use a `MINUS` construct, and we were trying to use only one with a UNION.

But since both exclusion statements are independent,
we can have two MINUS statements one after the other.

This removes a special case in `get_query_for_items_for_property_negative`,
and the dual `nega|postive_query_filter_out_fragment` in the Line objects,
which only existed for overloading in NoGroupGrouping.

Bug: T251008