Add Queries endpoint support for Property columns with qualifier
Support for qualifier columns (a special case of PropertyColumn)
was added all the way back to 6a055c2d47 (2019-10-25).

At the exact same time, we added the queries endpoint feature in 9e103b1d2f,
but it did not actually support qualifier columns.

Here we add support for the P1/P2 use-case ; the P1/Q2/P3
is likely not yet supported, but let's tackle that later.

Bug: T251008