Move grouping-link to Grouping objects and retrieve via SPARQL
Users can optionally format the grouping counts with an internal link,
where the target is the English label of the item.

Ever since ef5697d1, that value was retrieved
through the API (via `pywikibot.ItemPage`).

Here we instead retrieve the value via SPARQL.

This opens the door to allowing more
configuration for the grouping link.

It also has some nice benefits in terms of code:
- remove the dependency on the API ;
- remove need for the `repo` attribute in `PropertyStatistics`

Bug: T237276