Fix `QueriesTests` unit tests in test_app
The eight `QueriesTests` unit tests were failing with
```
unittest.mock.InvalidSpecError: Cannot autospec attr 'GROUP_MAPPING' as the patch target has already been mocked out.
```

The reason is that we were mocking
`pages_processor.PropertyStatistics.GROUP_MAPPING`,
but we already had a mock on `pages_processor.PropertyStatistics`

It is not clear when this started being a problem.
It appeared after upgrading pytest, but downgrading
pytest did not make the issue go away.