Fallback to mul when formatting grouping links
Users can optionally format the grouping counts with an internal link,
where the target is the English label of the item.

(The label is retrieved through SPARQL since 781934d).

Per the Wikidata documentation on `mul` [1]:
```
Items can be given default values for labels and aliases,
such as names and other proper nouns unlikely to differ between languages.
```
In some cases, the English label is being removed if
there is already a default label with the same value.

It makes sense to at least fallback to the
`mul` label if the `en` label is not present.

We should consider defaulting to `mul`,
but that’s part of the more general T237276.

[1] https://www.wikidata.org/wiki/Help:Default_values_for_labels_and_aliases

Bug: T393166