Optimize `get_grouping_information_query` query using subquery
Since 781934d, labels for groupings are retrieved
in SPARQL instead of using the MediaWiki API.

There were several good reasons for that move,
but one unexpected outcome is that it tanks
the performance of the grouping query:
some dashboards are now timing out.

Here we use a subquery to find the grouping and the count,
and then in the main query we retrieve the higher grouping
and the labels, which is more efficient.

This also remains standard SPARQL, so it is not
blocking potential support to other engine like Qlever.

Bug: T400480