Switch to login.toolforge.org as host for Ansible deployment
The Ansible playbook was initially configured in e2b2fcc
to target `login-stretch.tools.wmflabs.org`.

It was updated in 8bd793d to `login-buster.toolforge.org`
as part of the 2022 Toolforge Stretch deprecation [1].

Since then, the 2024 Buster deprecation happened [2],
and `login-buster` does not resolve anymore.

Let's switch to the plain `login.toolforge.org`,
which is the long-recommended form anyways [3]
(I’m not sure why I started with a specific bastion).

[1] https://wikitech.wikimedia.org/wiki/News/2022_Toolforge_Stretch_deprecation
[2] https://wikitech.wikimedia.org/wiki/News/2024_Buster_deprecation
[3] https://wikitech.wikimedia.org/wiki/Help:Toolforge/Quickstart#Connect_to_Toolforge_servers_using_SSH