Bump uv pre-commit hook from 0.5.24 to latest 0.9.5 and regenerate requirements files
We added a uv pre-commit hook in d13a7d966 (2025-01-24)
with then latest version 0.5.24 (2025-01-24) [1].

Let's upgrade to the most recent version 0.9.5 (2025-10-21) [2].

The default export format was changed 0.6.11,
with dependents added (as `via xyz` comments),
so this also regenerates the requirements files.

https://github.com/astral-sh/uv/releases/tag/0.6.11

The pre-commit passes, so the upgrade does not require any changes.

[1] https://github.com/astral-sh/uv/releases/tag/0.5.24
[2] https://github.com/astral-sh/uv/releases/tag/0.9.5
[3] https://github.com/astral-sh/uv/releases/tag/0.6.11