Simplify Flask webserver startup in app.py
The main method from the original 3178c6e5 commit was convoluted,
testing whether this it running on the Toolforge grid engine
and using a different runtime if so.

I cannot remember for sure where I got that boilerplate from.
I think I imported from the `wikiloves` repository, which had
also added it in its very first commit in 2016.

None of that is recommended anymore in the documentation [1]

[1] https://wikitech.wikimedia.org/wiki/Help:Toolforge/Web/Python