Implement SPARQL query engine dependency injection
Both `PropertyStatistics` and `GroupingConfiguration` make
SPARQL queries, using the `pywikibot.data.sparql` module.

Here we wrap the calls to pywikibot's module
(and thus the call to the SPARQL engine)
as the `WdqsSparqlQueryEngine` class.

The object can be injected into PropertyStatistics,
and it is also passed to `get_grouping_information`.

This requires quite a few changes to the PropertyStatistics unit tests.

This paves the way to use alternative SPARQL query engines (T385749).

Bug: T385749