Add support for QLever as alternative SPARQL query engine
QLever is an alternative SPARQL endpoint for Wikidata data,
considerably faster than the Blazegraph-based Wikidata Query Service.

Building on the asbtract introduced in 4da9f3338,
this adds support for QLever endpoint.

- Add all posible PREFIX declarations as required by QLever
- Transform response to be identical to the pywikibot one
- Add unit tests covering successful queries, timeout handling,
  and response transformation including real QLever API format.

For now this still defaults to WDQS and there is no interface to change.
This will be done in an upcoming commit.

Bug: T385749